pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sme-credit-model"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "pandas",
  "PyYAML",
  "openpyxl",
  "xlrd",
  "pydantic"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sme_credit*"]
exclude = ["tests*", "config*", "input_data*", "output_data*"]
